+++ cicd/utility/crosscheck.bash
@@ -227,6 +227,8 @@ fReadRow(){
 		              fCompare "fmt ${level:-standard}" fmt "${strictArg[@]}" "$input" ;;
 		count)        fCompare "count ${query}" count "${strictArg[@]}" "$input" "$query" ;;
 		instances)    fCompare "instances ${query}" instances "${strictArg[@]}" "$input" "$query" ;;
+		children)     fCompare "children ${query}" children "${strictArg[@]}" "$input" "$query" ;;
+		paths)        fCompare "paths" paths "${strictArg[@]}" "$input" ;;
 		*'[]')        fCompare "get ${query} ${type}" get "--${type%[]}" --array "${strictArg[@]}" "$input" "$query"
 		              fCompare "get ${query} ${type} slots" get "--${type%[]}" --array --slots "${strictArg[@]}" "$input" "$query" ;;
 		*)            fCompare "get ${query} ${type}" get "--${type}" "${strictArg[@]}" "$input" "$query"